chore: move getNextOpenText to schedule utils
